ai-code-reviewer/requirements.txt
streamlit==1.46.1
flake8==7.3.0
black==25.1.0
radon==6.0.1
bandit==1.8.6
google-generativeai==0.8.5
plotly==6.2.0